requirements.txt
streamlit
numpy
matplotlib
opencv-python
opencv-python opencv-contrib-python
pillow